ai-api/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name =  "voice-chatbot"
version = "0.0.1"
description = "Voice Chatbot API"
readme = "README.md"
requires-python = ">3"

dependencies = [
    "httpx==0.24.0",
    "jinja2==3.1.2",
    "pydantic==1.10.7",
    "urllib3==1.26.6",
    "google-api-python-client",
    "google-cloud-texttospeech",
    "fastapi",
    "uvicorn",
    "openai",
    "google.cloud",
    "google",
    "scipy",
    "click==8.1.3",
    "python-multipart",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff"
]

[project.urls]
Source = "https://github.com/marcorichetta/fastapi-vercel"

[tool.black]
line-length = 120
fast = true
